+++ sparrow-sharp/Sparrow/Geom/Matrix.cs
@@ -6,69 +6,69 @@ namespace Sparrow.Geom
 {
 	public class Matrix
 	{
-        // are you sure that these need to be double? BaseEffect needs a float Matrix
-		private double _a;
-		private double _b;
-		private double _c;
-		private double _d;
-		private double _tx;
-		private double _ty;
+        // are you sure that these need to be float? BaseEffect needs a float Matrix
+		private float _a;
+		private float _b;
+		private float _c;
+		private float _d;
+		private float _tx;
+		private float _ty;
 
-		public double A {
+		public float A {
 			get { return _a; }
 			set { _a = value; }
 		}
 
-		public double B {
+		public float B {
 			get { return _b; }
 			set { _b = value; }
 		}
 
-		public double C {
+		public float C {
 			get { return _c; }
 			set { _c = value; }
 		}
 
-		public double D {
+		public float D {
 			get { return _d; }
 			set { _d = value; }
 		}
 
-		public double Tx {
+		public float Tx {
 			get { return _tx; }
 			set { _tx = value; }
 		}
 
-		public double Ty {
+		public float Ty {
 			get { return _ty; }
 			set { _ty = value; }
 		}
 
-		public double Determinant {
+		public float Determinant {
 			get { return _a * _d - _c * _b; }
 		}
 
-		public double Rotation {
-			get { return Math.Atan2 (_b, _a); }
+		public float Rotation {
+			get { return (float) Math.Atan2 (_b, _a); }
 		}
 
-		public double ScaleX {
-			get { return _a / Math.Cos (SkewY); }
+		public float ScaleX {
+			get { return _a / (float) Math.Cos (SkewY); }
 		}
 
-		public double ScaleY {
-			get { return _d / Math.Cos (SkewX); }
+		public float ScaleY {
+			get { return _d / (float) Math.Cos (SkewX); }
 		}
 
-		public double SkewX {
-			get { return Math.Atan (-_c / _d); }
+		public float SkewX {
+			get { return (float) Math.Atan (-_c / _d); }
 		}
 
-		public double SkewY {
-			get { return Math.Atan (_b / _a); }
+		public float SkewY {
+			get { return (float) Math.Atan (_b / _a); }
 		}
 
-		public Matrix (double a = 1.0, double b = 0.0, double c = 0.0, double d = 1.0, double tx = 0.0, double ty = 0.0)
+		public Matrix (float a = 1.0f, float b = 0.0f, float c = 0.0f, float d = 1.0f, float tx = 0.0f, float ty = 0.0f)
 		{
 			_a = a;
 			_b = b;
@@ -98,35 +98,35 @@ namespace Sparrow.Geom
 			_ty = _ty * matrix.Ty + _d * matrix.Ty;
 		}
 
-		public void Translate (double dx, double dy)
+		public void Translate (float dx, float dy)
 		{
 			_tx += dx;
 			_ty += dy;
 		}
 
-		public void Scale (double sx, double sy)
+		public void Scale (float sx, float sy)
 		{
-			if (sx != 1.0) {
+			if (sx != 1.0f) {
 				_a *= sx;
 				_c *= sx;
 				_tx *= sx;
 			}
 
-			if (sy != 1.0) {
+			if (sy != 1.0f) {
 				_b *= sy;
 				_d *= sy;
 				_ty *= sy;
 			}
 		}
 
-		public void Rotate (double angle)
+		public void Rotate (float angle)
 		{
-			if (angle == 0.0) {
+			if (angle == 0.0f) {
 				return;
 			}
 
-			double cos = Math.Cos (angle);
-			double sin = Math.Sin (angle);
+			float cos = (float) Math.Cos (angle);
+			float sin = (float) Math.Sin (angle);
 
 			_a = _a * cos - _b * sin;  
 			_b = _a * sin + _b * cos;
@@ -136,12 +136,12 @@ namespace Sparrow.Geom
 			_ty = _tx * sin + _ty * cos;
 		}
 
-		public void Skew (double sx, double sy)
+		public void Skew (float sx, float sy)
 		{
-			double sinX = Math.Sin (sx);
-			double cosX = Math.Cos (sx);
-			double sinY = Math.Sin (sy);
-			double cosY = Math.Cos (sy);
+			float sinX = (float) Math.Sin (sx);
+			float cosX = (float) Math.Cos (sx);
+			float sinY = (float) Math.Sin (sy);
+			float cosY = (float) Math.Cos (sy);
 
 			_a = _a * cosY - _b * sinX;
 			_b = _a * sinY + _b * cosX;
@@ -153,12 +153,12 @@ namespace Sparrow.Geom
 
 		public void Identity ()
 		{
-			_a = 1.0;
-			_b = 0.0;
-			_c = 0.0;
-			_d = 1.0;
-			_tx = 0.0;
-			_ty = 0.0;
+			_a = 1.0f;
+			_b = 0.0f;
+			_c = 0.0f;
+			_d = 1.0f;
+			_tx = 0.0f;
+			_ty = 0.0f;
 		}
 
 		public Point TransformPoint (Point point)
@@ -166,14 +166,14 @@ namespace Sparrow.Geom
 			return new Point (_a * point.X + _c * point.Y + _tx, _b * point.X + _d * point.Y + _ty);
 		}
 
-		public Point TransformPoint (double x, double y)
+		public Point TransformPoint (float x, float y)
 		{
 			return new Point (_a * x + _c * y + _tx, _b * x + _d * y + _ty);
 		}
 
 		public void Invert ()
 		{
-			double det = Determinant;
+			float det = Determinant;
 
 			_a = _d / det;
 			_b = -_b / det;
