+++ sparrow-sharp/Sparrow/Utils/VertexColor.cs
@@ -4,9 +4,9 @@ namespace Sparrow.Utils
 {
 	public class VertexColor
 	{
-		public uint R;
-		public uint G;
-		public uint B;
-		public uint A;
+		public byte R;
+		public byte G;
+		public byte B;
+		public byte A;
 	}
 }
